+++ lib/node_program_editor.dart
@@ -1,1 +1,3 @@
-
+export './widgets/canvas.dart';
+export './models/node_data_model.dart';
+export './models/port_model.dart';